refactor : 하나의 controller를 user, game controller로 분리
